Fixed 'transparent' background color of tray icon
Replaced eggtrayicon code with version from revision r6499 of the "uim" project to provide for transparent tray icon backgrounds (with minor adjustments to #include and #define statements to make it compatible with this project).
http://code.google.com/p/uim/source/detail?r=6499&path=/trunk/helper/eggtrayicon.c

Note:  This doesn't seem to work in IceWM, however it does work in JWM.  One unfortunate catch is that it seems to require restarting X for the new background color to kick in.  Not yet sure why.